move pkg/controller/route to k8s.io/cloud-provider/controllers/route for easier external consumption
