+++ hack/.golint_failures
@@ -77,7 +77,6 @@ pkg/controller/replication
 pkg/controller/replication/config/v1alpha1
 pkg/controller/resourcequota
 pkg/controller/resourcequota/config/v1alpha1
-pkg/controller/route
 pkg/controller/service/config/v1alpha1
 pkg/controller/serviceaccount/config/v1alpha1
 pkg/controller/statefulset
@@ -424,6 +423,7 @@ staging/src/k8s.io/client-go/util/retry
 staging/src/k8s.io/client-go/util/testing
 staging/src/k8s.io/client-go/util/workqueue
 staging/src/k8s.io/cloud-provider
+staging/src/k8s.io/cloud-provider/controllers/route
 staging/src/k8s.io/code-generator/cmd/client-gen/args
 staging/src/k8s.io/code-generator/cmd/client-gen/generators/fake
 staging/src/k8s.io/code-generator/cmd/client-gen/generators/scheme
